backend/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
pytesseract
pillow
python-multipart
typing-extensions